eco_refs_t

+++ resources/views/ecorefsavgmarketprice/edit.blade.php
@@ -28,7 +28,7 @@
                                 <div class="col-xs-12 col-sm-12 col-md-12">
                                     <div class="form-group">
                                         <strong>Ставка ЭТП, $ за 1 тонну:</strong>
-                                        <input type="integer" name="exp_cust_duty_rate" class="form-control" placeholder="" value="{{$EcoRefsAvgMarketPrice->exp_cust_duty_rate}}">
+                                        <input type="float" name="exp_cust_duty_rate" class="form-control" placeholder="" value="{{$EcoRefsAvgMarketPrice->exp_cust_duty_rate}}">
                                     </div>
                                 </div>
                                 <div class="col-xs-12 col-sm-12 col-md-12 text-center">
